+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './tableGridColumn/tableGridColumn.component';
+export * from './tableGridMetadataGatherer/tableGridMetadataGatherer.component';
+export * from './tableGridMetadataGatherer/tableGridMetadataGatherer.interface';