+++ src/directives/index.ts
@@ -1,0 +1,4 @@
+export * from './tableGridBodyCellTemplate/tableGridBodyCellTemplate.directive';
+export * from './tableGridCellTemplate/tableGridCellTemplate.context';
+export * from './tableGridCellTemplate/tableGridCellTemplate.directive';
+export * from './tableGridHeaderCellTemplate/tableGridHeaderCellTemplate.directive';