+++ src/interfaces/gridColumn/index.ts
@@ -1,3 +1,2 @@
 export * from './gridColumn.interface';
 export * from './tableGridColumn.interface';
-export * from './tableGridColumnTemplate.context';