+++ src/interfaces/gridColumn/tableGridColumn.interface.ts
@@ -1,7 +1,7 @@
 import {TemplateRef} from '@angular/core';
 
 import {GridColumn} from './gridColumn.interface';
-import {TableGridColumnTemplateContext} from './tableGridColumnTemplate.context';
+import {TableGridCellTemplateContext} from '../../directives/tableGridCellTemplate/tableGridCellTemplate.context';
 
 /**
  * Table grid column definition
@@ -11,45 +11,45 @@ export interface TableGridColumn<TData = unknown> extends GridColumn
     /**
      * Name of property which is assigned to this column, can use . notation
      */
-    name: string|undefined|null;
+    readonly name: string|undefined|null;
 
     /**
      * Text that is displayed in tooltip over grid header
      */
-    headerTooltip: string|undefined|null;
+    readonly headerTooltip: string|undefined|null;
 
     /**
      * Indication whether should be title visible in header
      */
-    titleVisible: boolean|undefined|null;
+    readonly titleVisible: boolean|undefined|null;
 
     /**
      * Width as style string, value is exactly same (require units)
      */
-    width: string|undefined|null;
+    readonly width: string|undefined|null;
 
     /**
      * Indication that this column can be used for ordering
      */
-    ordering: boolean|undefined|null;
+    readonly ordering: boolean|undefined|null;
 
     /**
      * Css class that is applied to column header
      */
-    headerClass: string|undefined|null;
+    readonly headerClass: string|undefined|null;
 
     /**
      * Css class that is applied to each column cell
      */
-    cellClass: string|undefined|null;
+    readonly cellClass: string|undefined|null;
 
     /**
      * Template that is used for rendering of cell header
      */
-    headerTemplate: TemplateRef<TableGridColumnTemplateContext<TData>>|undefined|null;
+    readonly headerTemplate: TemplateRef<TableGridCellTemplateContext<TData>>|undefined|null;
 
     /**
      * Template that is used for rendering of cell body
      */
-    bodyTemplate: TemplateRef<TableGridColumnTemplateContext<TData>>|undefined|null;
+    readonly bodyTemplate: TemplateRef<TableGridCellTemplateContext<TData>>|undefined|null;
 }